+++ src/con_linux.cpp
@@ -1120,8 +1120,8 @@ int GUI::SetPipeView(int id, EModel *notify) {
     return 0;
 }
 
-int GUI::ReadPipe(int id, void *buffer, int len) {
-    int rc;
+ssize_t GUI::ReadPipe(int id, void *buffer, int len) {
+    ssize_t rc;
 
     if (id < 0 || id > MAX_PIPES)
         return -1;
