+++ src/con_os2.cpp
@@ -1063,8 +1063,8 @@ int GUI::SetPipeView(int id, EModel *notify) {
     return 0;
 }
 
-int GUI::ReadPipe(int id, void *buffer, int len) {
-    int l;
+ssize_t GUI::ReadPipe(int id, void *buffer, int len) {
+    ssize_t l;
     //ULONG ulPostCount;
 
     if (id < 0 || id > MAX_PIPES)
